webkit2.navigation-action.lisp: support creation of WebKitNavigationAction
we need to define an opaque g-boxed type to handle automatic creation
in g-signal callbacks

said another way:

we define-g-boxed-opaque for webkit-navigation-actions
because cl-cffi-gtk automatically attempts to create gobjects from
callbacks. We do not readily use the opaque gobject, for our
purposes it is no different than a pointer, so we use
(gobject:pointer webkit-navigation-gobject) to get and use the
pointer as necessary.